SVY-17070 [Ti] Calendar picker becomes useless after inputing invalid date time
updated our build of the datepicker to the latest release of the
original one. (or latest source)
implemented the hooks a bit different, more in how you now should plugin
stuff. Did make a pull request to make this nicer/better

fix the imp so that "Invalid Date" Date objects are never returned but
then just null (in 2 places)

make sure to test for the NULL_VALUE object of the datagrid (because
that value could be given instead of a null object)
We expect that the value is null or a date.